+++ src/DetectorImage.jsx
