+++ src/DetectorImageAll.jsx
