+++ src/DetectorSelector.jsx
